refactor: move shouldCheckpointAfterSync into checkpointDecision module
